app/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
Deprecated==1.2.14
fastapi==0.103.2
gcloud==0.18.3
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
oauth2client==4.1.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
Pyrebase4==4.7.1
python-jwt==4.0.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
wrapt==1.15.0
